WebUI: Enforce single naming convention for manifest files.
Prior to this change, some targets output manifests to *.manifest, while
others output to *_manifest.json. This CL conforms all usages
to the latter convention. Notably, ts_library() was changed to output
manifests to ${target_name}_manifest.json by default.

This is the better convention as it preserves the filetype. This has
benefits in the context of highlighting in editors as well as improving
code clarity.

Change-Id: I67509fbc0f4e7caddf2cbbdf905aa3ac17710e59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4562723
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Cole Horvitz <colehorvitz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150707}