Rework Graphics.Smoothness.Checkerboarding to use V3 signals
For PercentDroppedFrames we upgrading from V1 to V3 with more accurate
signals for the importance of the frames. V3 is now a guardrail metric.

We want to eliminate the old V1 path. However some metrics still use it.
Checkerboarding is one such metric. This change updated V3 to perform
the calculation.

Bug: 1256879
Change-Id: I99873bfe825e849583faafbec5e1700d126fefc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4573781
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Mingjing Zhang <mjzhang@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150727}