Fix NOTREACHED when setting automatic downloads to ALLOW_ALL_DOWNLOADS
I noticed a bug within the automatic downloads flow: when the setting
is changed to ALLOW_ALL_DOWNLOADS via the dialog, the icon remains in
the omnibox. However, it shows an empty dialog when clicked. This
happens because we're hitting the NOTREACHED in
ContentSettingDownloadsBubbleModel::SetRadioGroup(). This is causing
crashes in Edge, hence the need to fix.

This CL makes sure that we don't erase the status in
OnUserInteraction() when it is ALLOW_ALL_DOWNLOADS. We reset the status
until DidFinishNavigation(). This renders a dialog with the options to
continue accepting or block automatic downloads.

Repro steps:
1) Navigate to a website with an automatic download.
2) Set automatic downloads to "Don't allow".
3) Try to download, it is blocked.
4) Select "Always allow" in dialog.
5) Download file.
6) Open dialog, see that content is empty.

Change-Id: I242fec7045a07e1f5fc995bf711a5d819369c1c7
Bug: 1423334
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4534427
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Amanda Velasco Gallardo <amandave@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1150812}