image_processor_perftest: Implement Uncapped Test for LibYUV
Previously the image_processor_perftest was created with an uncapped
performance test for the GLImageProcessor but not for LibYUV.

Change adds an uncapped test for LibYUV to the image_processor_perftest
suite. The test setup and format are exactly the same as in the Uncapped
GLImageProcessor test with the only difference being that the frames are
being processed by LibYUV instead.

the Uncapped LibYUV test.

BUG: b:278107121
Test: Tested on an Asurada board and was able to get test results from
Change-Id: Ib4af1536821c4f264261938643f5538de495951e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4562520
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Justin Green <greenjustin@google.com>
Cr-Commit-Position: refs/heads/main@{#1150944}