saved_desks: Fix excessive repainting in the library
This CL fixes excessive repainting in the saved desk library. Prior to
this CL, just moving the cursor around in the saved desk library would
cause the entire library to be repainted.

If the cursor is moved to hover over a saved desk item, we fade in the
button to launch the template. It is positioned in the same spot as the
icons for the saved desk, so those fade out. In the old implementation,
once the icons had faded out, we would call SetVisibility on the icon
container view. This in turn triggered the containing flex layout, which
eventually results in a lot of stuff getting laid out again, which then
ends up re-painting things.

This change makes it so that the two views (the icon container and the
hover container) are always "visible". We instead modify the opacity on
the containers as needed to hide and show them. With this change in
place, no `Layout` calls happen when moving the cursor around.

BUG=b:281880456

Change-Id: Iea5bbd494ae308b81f597081ccb50db111c917df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4562379
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150962}