jinx: fix prediction suggestion tracking
Currently tracking of prediction suggestions is incorrect when the first
character of a suggestion is not typed by the user. This is due to a bug
in the current tracking logic where the MultiWordSuggester does not
enter tracking mode when a prediction suggestion is initially shown.

This CL fixes this logic with the new stale suggestion handling logic
added in crrev/c/4520736. With stale suggestions handled, we then know
that any prediction suggestion received should simply be appended to the
current surrounding text. Thus we can transition the suggester straight
to tracking mode without needing to reconcile the current surrounding
text with the suggestion given.

BUG=b:278800045

Change-Id: I5fc5db24fb7a01eaa7fc14371d698fabd2d28423
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520080
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Mehrab N <mehrab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150968}