[ios] Fix Fake Omnibox bug when sticking to top of NTP
The logic in NTPViewController's -offsetToStickOmnibox to determine
when to stick the fake omnibox to the top of the surface is not the
same as the one within the NewTabPageView that determines when the
fake omnibox expansion animation stops (which happens to be the same
position as where it should stick to the top of surface). This change
makes NewTabPageView's logic public to the NTPViewController.

Video: https://drive.google.com/file/d/1tmjMxIrzQjah1wlUzXfrerpYV-M8TKV2/view?usp=sharing

Fixed: 1443603
Change-Id: I0aa81e8aceb77cae41068b08db70be08e2d2bda2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4569060
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1150978}