[text-wrap-pretty] Support block-in-inline
This patch supports block-in-inline for `text-wrap: pretty`.

Due to an performance optimization in `HandleBlockInInline`,
`NGLineBreaker` needs the same `NGExclusionSpace` as the one
in the `constraint_space_`. This patch adds it to the
`NGScoreLineBreaker` constructor.

`NGScoreLineBreaker` computes the line breaking by splitting
the block to 3 parts; before the block-in-inline, the block-
in-inline itself, and after, as if the 3 parts are separate
paragraphs. The unit tests cover this state transitions.

Bug: 1432798
Change-Id: I7df609418f08839a422ec65f28f36b5d840619e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4565457
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150989}