[SPEAK-ON-MUTE] Add controller to privacy hub
Adds an extra controller to set the speak-on-mute state when the user
logins and monitors the speak-on-mute detection enabled pref changes.

The detection state is no longer set during the cras_audio_handler
initialization, but when the user logs in instead. Because it is a
per-profile feature instead of per-device as discussed.

gtest_filter='*PrivacyHubSpeakOnMuteControllerTest*'

Bug: b/280925839, b/283345021
Test: Manually test on DUT
Test: chromeos_unittests --gtest_filter='*CrasAudioHandlerTest*'
Test: ash_unittests --
Change-Id: I29f8e1d6c1dba8b9b12754ae667ac0db97011e13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4563594
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: CJ Huang <chenjih@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Charles Zhao <charleszhao@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Cr-Commit-Position: refs/heads/main@{#1151007}