sql: Add wrapper to facilitate migration to the new recovery module
This new method decides internally whether to use the new
`sql::BuiltInRecovery` or the legacy `sql::Recovery`, based on
feature flags and availability. It is intended to facilitate the
migration to the newer recovery module with minimal impact on feature
teams. The expectation is that the branch which uses `sql::Recovery`
will eventually become obsolete and be removed entirely.

Adds a new kill switch which may disable all use of `BuiltInRecovery`
if necessary.

Also updates the test suites to remove use of `if (!BuiltIn()) ...`

Bug: 1385500
Change-Id: Idaeab746c7a257b1eda2f1ec02c6525c7473bcdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4570917
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151027}