Fix OpenUrlsInBrowserWithProfile to prefer a regular profile
Before this change, on macOS, if an incognito window was opened as the
frontmost window, a url from other app (e.g. Mail app) was opened on the
incognito window. The reason behind this erroneous behavior was that
`OpenUrlsInBrowserWithProfile` used the passed profile untouched even if
the passed profile was an incognito profile.

This change fixes `OpenUrlsInBrowserWithProfile` to prefer a regular
(non-incognito) profile.

This change also enables a related browser test and fixes its flakiness:
AppControllerBrowserTest.OpenInRegularBrowser

Fixed: 1444747, 1354858
Bug: 757253
Include-Ci-Only-Tests: true
Change-Id: I21db42e2432ae54ccc9ce56bde961f62b65e75f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4530255
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151116}