BrowserLoader: Refine background stateful loading log
This CL updates the log on loading stateful lacros in the background.
In case where stateful lacros loading failed in the background, it
looked strange with the existing code:
"stateful lacros-chrome installation completed in the background in ,
version is invalid"
Refined it to look better.

This CL does not affect the behavior.

Bug: b:270164115
Change-Id: Ic9441f62eb1817214742dbbe6f19304c476c4264
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4570315
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151210}