[//viz] s/plane/mailbox in entities involved in CopyOutputNV12() flow
The flow around CopyOutputNV12() currently makes the assumption that a
single mailbox == a single plane throughout. This is reflected in the
usage of "plane" and "Plane" in the naming throughout this flow.

In preparation for introducing handling of NV12_MULTIPLANE in this flow,
this CL performs renames to decouple "mailbox" from "plane". This will
allow for smoothly introducing a variant of the flow that processes one
mailbox for all of the planes. The renames are the following:

PlaneAccessData -> MailboxAccessData
plane_access_datas -> mailbox_access_datas
NV12SinglePlaneReadyContext -> NV12SingleMailboxReadyContext
plane_mailbox_holder -> mailbox_holder
plane_data -> mailbox_data
OnSinglePlaneReady() -> OnSingleMailboxReady()

There are no functional changes in this CL.

Bug: 1429004
Change-Id: Ieb9080dae62f0e1a0b44490c805a0cd33dfafeb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4571691
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Colin Blundell <blundell@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151246}