Re-order member destruction in ~ComponentUpdaterTest()
Avoid a spurious dangling ptr warning by ensuring that `scheduler_`
comes after `component_updater_` so that it outlives it when destroyed
in inverse order of declaration. Then remove the explicit reset()
call in the destructor and rely on in-order destruction.

Then do the same for ComponentInstallerTest().

Bug: 1401495
Change-Id: Ie3c944fda2e4b83f92e36ee6c6e5a43e9bfbaa47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575411
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151257}