Include ad render Id in prevWins metadata
We also change the way we're storing prevWins metadata. Previously we
assembled the full JSON payload with a series of sprintfs. Now we keep
the metadata serialized and only deserialize it for generateBid. This
is useful since it means we don't have to deserialize the metadata when
looking for the adRenderId used in B&A style auctions.

Bug: 1442274
Change-Id: If1fa8157c2ccea382607f9f9c7ea24784da4011b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4566565
Commit-Queue: Russ Hamilton <behamilton@google.com>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151258}