Icons on parents of submenus should be included in the icon max width.
For example, bookmark folders are a submenu, so none of their icons were
included in this calculation.

The effect of this was severe, as the overall menu width and the text
start locations are computed based on this value, so wrongly setting it
to zero would result in very glitchy-looking menus. This is more visible
than it used to be since r1148577 stopped reserving the check width even
for menus without checks/radios.

This also does a little more cleanup -- be consistent about whether to
use has_icons_ or has_icons(), and don't null-check GetRootMenuItem()
(which can't return null).

Bug: none
Change-Id: Ic98874b5167e4fcacd58dfe3897e8fbb5f00a8c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4576162
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151269}