Hide detail label from continue chips if empty
When querying for SearchResults for the continue section, we used to get
a field for details on the file. Recently, some issues with locale
caused the drive suggestions to provide an empty string in details. As a
result, the label would not display any text, but take the space
assigned to it, causing the title to looks not centered across the
vertical axis.

As a mitigation for this, if the chip has an empty details field, hide
the details label for the LayoutManager to center the title label in the
vertical axis.

Bug: b:279644935
Change-Id: I1ecb81e998c587693a2c156222dd8d362611867b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575448
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151286}