[A11y] Prepare to share some extension code with Lacros
Select to Speak and Switch Access both inject a content script into
Google Docs that informs Docs to annotate the canvas for a11y. This
script was not injected in Lacros browsers because STS/Switch run as
component extension in Ash.

This change prepares some code in Chrome OS extensions to share with
Lacros in an upcoming patch: specifically manifest generation template
and the gdocs_script.

Test:Manual from crrev.com/c/4569470, compile.

Ax-Relnotes: N/A
Bug: b:272150598
Change-Id: I03de3e9a1920e0d04ce4193422197bd85b7893d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575567
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151313}