[Auto-disable Accessibility] Remove forced announcements from Feed in favor of live region
This CL continues the work of consolidating accessibility management to
a single source in Clank.

With this CL we remove the hard dependency on AccessibilityManager from
the FeedSwipeRefreshLayout. Rather than relying on AccessibilityState to
fetch the same information, we replace the logic using a live region.
This is the preferred method per the API and Google's GAR. Disruptive
announcements should be avoided and must be approved by the awg@, see
more info at: go/noannounce and GAR criterion 1.3:
go/GAR-mobile#disruptive-announcements

By using the live region we are also providing the information to all
AccessibilityServices that may be interested in it, rather than making
an announcement that is TalkBack specific.


AX-Relnotes: N/A
Bug: 1430202, b/265493191
Change-Id: I2b88f1078c815d9a0670981ede95d665c34bfb4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575327
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jian Li <jianli@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1151325}