[iOS] Migrate OverflowMenu ranking to new pref and prevent wiping it
There was a bug where the DestinationUsageHistory would wipe and re-
write the kOverflowMenuDestinationUsageHistory pref without filling
in the ranking, and the OverflowMenuOrderer might not get a chance to
add the ranking list back in.

This solves the issue in two ways. First, DestinationUsageHistory
carries over any stored ranking in the pref when it's writing prefs
back out. Second, OverflowMenuOrderer migrates the ranking to a new
separate pref key.

Bug: 1448271
Change-Id: Ib811a3c96771e6d0a1d39c4b7699458d6fe1fc05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4557210
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1151333}