[DownloadBubble] Temporarily show the toolbar in immersive fullscreen
This makes the download bubble temporarily reveal the toolbar when it is
shown in immersive fullscreen mode on ChromeOS and Mac. On other
platforms and in non-immersive fullscreen, the behavior is unchanged
(there appears an "exclusive access bubble"/banner telling the user to
exit fullscreen to view their downloads).

Note that we cannot simply call BrowserView::FocusToolbar() because this
actually focuses the toolbar and the focus remains on the toolbar when
the download bubble closes, which we don't want. Thus, we need to
manually take out an ImmersiveRevealedLock so that the toolbar will
automatically hide when the download bubble closes.

This also fixes a bug where we were showing an exclusive access bubble
telling the user to exit fullscreen to view their downloads, when it was
impossible to exit fullscreen, such as in kiosk mode. And renames a bool
to a more accurate name.

Screencasts:
https://drive.google.com/drive/folders/15TyjGZ1-zxLZBg2sE2MVn6YLJiO0PC3I

Bug: 1444516, 1416640
Change-Id: Iba3305b06b4b3714c428147a1c64c5b795517693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4539802
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151350}