Remove LoginState::SetLoggedInStateAndPrimaryUser.
Primary user hash can be obtainted from global
user_manager::UserManager.
This CL removes the cache in LoginState for reducing the dependency,
while the method is kept for the compatibility for the short term.
The method will be removed with a few more following CLs soon.

BUG=b:27864311
TEST=Tryjob.

Change-Id: I4ea4de010c2cab8d8259f2b4a10ea16859e34399
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4543780
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151368}