[Android] Use FeatureMap to generalize ContentFeatureList
This reduces code duplication in ContentFeatureList.java and
content_feature_list.cc, reusing the code in base
FeatureMap.java and feature_map.cc.

ContentFeatureList is kept for convenience static methods to
access ContentFeatureMap, but we can inline and update callers in
a follow-up if desired.

Also add a TODO to ContentFeatureList since I noticed the constants
there with feature names largely overlap with the constants in the
generated ContentFeatures.

Using this opportunity to update documentation
(since ContentFeatureList is used as an example) with the new
FeatureMap API.

Change-Id: I2360830bfbceff20714fc65db4dce5ef743ab95a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4544343
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151372}