Allow generating the stdlib GN rules from //third_party/rust_src
This allows the GN rules to be generated before a Rust roll completes,
since the roll requires the new GN rules in order to succeed.

To do so, pass a directory relative to the chromium source tree as
a value to `--for-std` such as `--for-std=third_party/rust_src/src`
and then the Cargo rules and vendored packages will be used from
that directory instead of the default which is the installed
toolchain in //third_party/rust-toolchain.

Bug: 1401042
Change-Id: Id1515295eb8002d2e4fb180a04ae6aad234012e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4569058
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151374}