Replacing 11 boolean typed metrics into one bit vector of an enum list
We are working on saving the UKM storage size of Autofill2.FieldInfo,
and we are thinking of merging 11 boolean typed metrics into one
number, which is a bit vector of an enum of 12 values. Please see the
discussion here: https://buganizer.corp.google.com/issues/280629804.

A metric of a boolean type is 1 byte, and a bit vector of an enum of 12
values is 2 bytes, so we save 11-2=9 bytes for each record of
Autofill2.FieldInfo.

In order to make the data consistent, we are keeping both the boolean
metrics of autofill status and this new bit vector for now. Once we
verify the data in AutofillStatusVector metric, we will mark the
according boolean metrics obsolete.

Bug: 1325851
Change-Id: Id8b1425dc07a06f7a73a5116138cbed58d9a8942
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4517979
Reviewed-by: Sun Yueru <yrsun@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151376}