Speak On Mute Opt-in nudge
Implement the anchored nudge for users to opt in or opt out of the
Speak On Mute Detection feature. The nudge attempts to be shown every
time the VC tray visibility changes to true and the mic is muted or the
mic was muted and the tray was visible.

screenshot: https://screenshot.googleplex.com/5WuAUJpeyX6y4nu

The nudge will not show again if:
- The user interacted with the nudge before.
- The user toggled on the Speak On Mute feature through settings.
- The nudge was shown its max amount of times.

Bug: b:274519463
Change-Id: Idc29f6cb4e60ad3cf4d34adc4846efd9963aea0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4549238
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Kevin Radtke <kradtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151390}