Fixes issue where min framerate might not be fulfilled for screensharing
The current RequestRefreshFrame (RRF) scheme is very restrictive when
it comes to sending a RRF frame and heavily relies on the using the
default capture stream instead.

It works in normal cases where the client only asks for a refresh at
random (rare) cases, e.g. when starting a stream etc.

However, the current scheme breaks an important case when using
applyConstraints in combination with getDisplayMedia to specify min
and max frame rate. If e.g using min=max=30 (fps), the client expects
to get a *minimum* frame rate of 30 fps. But that is not the case today
if the captured source is static since the default capture stream will
then not send any new frames unless the source has changed. And all
attempts from the client to ask for a refresh will be ignored since
the VDC thinks "I am about to send a default frame soon, hence no need
for an RRF". The only state which today triggers an RRF is if the
default captured frame is very delayed and that is rare.

The new scheme removes the old RRF timer and instead "increases the
priority of refresh frames" by simply converting them into default
capture calls with one important difference; they also ignore the 0Hz
mode and ensures that the captured frame is *always* sent to the
client. This way it is now possible for the client to ask for a certain framerate even for static content. It is not power efficient but it follows the spec.

PS1 This change avoids holding on to a copy of the last captured
frame for potential use as RRF since we know that this can cause
severe cursor flickering.

PS2 I plan to add a new unittest as a follow-up but would like to
fix this ASAP since we break existing clients as is.

Bug: 1421656
Change-Id: Ifa1870dd93c123f2d5c7f01e7790a412d0ec602d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4568018
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Henrik Andreasson <henrika@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151410}